requirements.txt
PySimpleGUI==4.60.5
pytesseract==0.3.10
Pillow==10.4.0
PyMuPDF==1.24.5
pywin32==306
